refactor: refatora código dos testes sem mudança de lógica
Separa o código dos testes em funções para futuro uso em outros testes
