+++ frontend/src/tests/TaskForm.test.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import '@testing-library/jest-dom';
 import { screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import customRender from './utils/customRender';
+import { customRender, multipleTestWrapper, generateTestMessage } from './utils';
 
 import TaskForm from '../Components/TaskForm';
-import { getAllTasks, getRandomTask } from './mock/getRandomTask';
+import { getAllTasks } from './mock/getRandomTask';
 
 describe('Testa o componente TaskForm', () => {
   describe('recebe mensagem de erro quando não insere os dados incorretos', async () => {
@@ -13,21 +12,11 @@ describe('Testa o componente TaskForm', () => {
 
     const tasks = getAllTasks('invalid');
 
-    const inputInfo = screen.queryByRole('textbox');
-    const inputStatus = screen.queryByTestId('form-task-status');
-    const addButton = screen.queryByRole('button', { name: /adicionar tarefa/i });
-
     tasks.forEach((task) => {
-      it(`testa caso: ${JSON.stringify(task, null, '')}`, () => {
+      it(generateTestMessage(task), () => {
         const warningDuration = 2001;
-        setTimeout(() => {
-          userEvent.type(inputInfo, task.info);
-          userEvent.selectOptions(inputStatus, task.status);
-
-          expect(inputInfo).toHaveValue(task.info);
-          expect(inputStatus).toHaveValue(task.status);
-
-          userEvent.click(addButton);
+        multipleTestWrapper(() => {
+          const [inputInfo, inputStatus] = makeNewTask(task);
 
           expect(inputInfo).toHaveValue(task.info);
           expect(inputStatus).toHaveValue(task.status);
@@ -40,28 +29,20 @@ describe('Testa o componente TaskForm', () => {
     });
   });
 
-  it('adiciona tarefa com sucesso', async () => {
+  describe('adiciona tarefa com sucesso com dados corretos', async () => {
     customRender(<TaskForm edit="" />);
 
-    const task = getRandomTask('valid');
+    const tasks = getAllTasks('valid');
 
-    const inputInfo = screen.queryByRole('textbox');
-    const inputStatus = screen.queryByTestId('form-task-status');
-    const addButton = screen.queryByRole('button', { name: /adicionar tarefa/i });
+    tasks.forEach((task) => {
+      it(generateTestMessage(task), () => {
+        multipleTestWrapper(() => {
+          const [inputInfo, inputStatus] = makeNewTask(task);
 
-    userEvent.type(inputInfo, task.info);
-    userEvent.selectOptions(inputStatus, task.status);
-
-    expect(inputInfo).toHaveValue(task.info);
-    expect(inputStatus).toHaveValue(task.status);
-
-    userEvent.click(addButton);
-
-    const timeout = 500;
-
-    setTimeout(() => {
-      expect(inputInfo).toHaveValue('');
-      expect(inputInfo).toHaveValue('');
-    }, timeout);
+          expect(inputInfo).toHaveValue('');
+          expect(inputStatus).toHaveValue('');
+        });
+      });
+    });
   });
 });
