+++ frontend/src/tests/actions/index.js
@@ -1,0 +1,7 @@
+import makeNewTask from './makeNewTask';
+
+export default makeNewTask;
+
+// export {
+//   makeNewTask,
+// };
