+++ frontend/src/tests/utils/generateTestMessage.js
@@ -1,0 +1,5 @@
+function generateTestMessage(object) {
+  return `test: ${JSON.stringify(object, null, '')}`;
+}
+
+export default generateTestMessage;
