+++ frontend/src/tests/utils/index.js
@@ -1,0 +1,9 @@
+import multipleTestWrapper from './multipleTestWrapper';
+import customRender from './customRender';
+import generateTestMessage from './generateTestMessage';
+
+export {
+  multipleTestWrapper,
+  customRender,
+  generateTestMessage,
+};
