+++ frontend/src/tests/utils/multipleTestWrapper.js
@@ -1,0 +1,7 @@
+function multipleTestWrapper(test, timeout = 1) {
+  setTimeout(() => {
+    test();
+  }, timeout);
+}
+
+export default multipleTestWrapper;
